Refactors socrata_tasks.py to use the fully task-grouped implementation and adapts two update DAGs.
